# Dokumentáció/Fejlesztői napló/Api/pipeline_swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "PipeLine.Backend",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Device/myFilteredDevices": {
      "get": {
        "tags": [
          "Device"
        ],
        "parameters": [
          {
            "name": "input",
            "in": "query",
            "schema": {
              "type": "string",
              "default": ""
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Device/FilteredDevices": {
      "get": {
        "tags": [
          "Device"
        ],
        "parameters": [
          {
            "name": "input",
            "in": "query",
            "schema": {
              "type": "string",
              "default": ""
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Device": {
      "post": {
        "tags": [
          "Device"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "tags": [
          "Device"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Device"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Device/{id}": {
      "get": {
        "tags": [
          "Device"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Device"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeviceDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "deviceType": {
            "$ref": "#/components/schemas/DeviceType"
          },
          "manufacturer": {
            "type": "string",
            "nullable": true
          },
          "model": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "applicationUserId": {
            "type": "string",
            "format": "uuid"
          },
          "batteryCapacity": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "maxChargingSpeed": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "batteryVoltage": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "detachableBattery": {
            "type": "boolean",
            "nullable": true
          },
          "isFoldable": {
            "type": "boolean",
            "nullable": true
          },
          "canBeLocked": {
            "type": "boolean",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DeviceType": {
        "enum": [
          0,
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
...
    }
  }
}